+++ store/tikv/oracle/oracles/mock.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mockoracle
+package oracles
 
 import (
 	"context"
